Inheritance example
To understand the concepts of inheritance.
here the important thing is to be noted is specifier -- protected is used bcoz by using this we can access the properties inside the package and the sub classes, if we use public it will accessible outside the class and if we use private it will accessible only inside the class. So to avoid these extremes we use protected specifier.