practicing method overriding
To practice method overriding and to understand dynamic method dispatch and to use annotation.
Annotations in Java provide additional information to the compiler and JVM. An annotation is a tag representing metadata about classes, interfaces, variables, methods, or fields. Annotations do not impact the execution of the code that they annotate.