Abstract class
Using abstract keyword we made the abs cls which is nothing but incomplete cls, that is it has 0 or 1 or more abs methods in it. abs method is nothing but incomplete method without body.
so for we cant create obj for abs cls bcoz its incomplete but we can make refernce of abs cls. we can inherit a cls from abs cls and that is also abs cls., to make it concrete cls v hv to override the methods which present inside the abs cls and dont use abs keyword in inherited cls.