exception handling in stack
to understand user defined exception class and the exception handling and try catch block along with the concept called stack. Where stack is a data structure and it follows last in first out principle, it follows array so size is fixed. There are two methods push and pop for insert and delete elements,. and the exceptions are stack overflow and underflow.